Add module for function toResponse which handles mongo _id
